Add files via upload
Abstract:
Instrumental Variable (IV) estimators have generally been used to lower the bias from Ordinary Least Squares (OLS) when performing a regression, due to the fact that IV accounts for endogeneity problems such as omitted variable bias. However, in certain circumstances, it is possible for IV estimators to be more biased than the OLS, even when endogeneity is present. This research will be aimed at presenting those situations in which OLS outperforms IV estimates, through iteratively manipulating variances for the instrumental variable and the endogenous regressor. Further, we determine instances in which the exclusion condition of the IV technique (that the error term is uncorrelated with the instrumental variable) is more important than the inclusion restriction (that there is a correlation between the endogenous regressor and the instrumental variable).